Only delete keywords, not entire item
This fixes a bug where removing all tags from a file would also remove it from any collections you'd added it to, and safeguards against similar bugs if Ableton add extra metadata later on.

A slightly better approach would be to check if the item only contains 'filePath' before removing the item - I'll try to add that later on.